sales/wws/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
pdf2image==1.16.3
mysql-connector-python==8.2.0
Pillow==10.1.0
schedule==1.2.0
selenium==4.15.2
webdriver-manager==4.0.1